+++ RealmSwift/Object.swift
@@ -81,7 +81,7 @@ import Realm.Private
  See our [Swift guide](https://docs.mongodb.com/realm/sdk/swift/fundamentals/relationships/) for more details.
  */
 public typealias Object = RealmSwiftObject
-extension Object: RealmCollectionValue {
+extension Object: _RealmCollectionValueInsideOptional {
     // MARK: Initializers
 
     /**