+++ RealmSwift/Tests/TestValueFactory.swift
@@ -47,25 +47,16 @@ struct UnmanagedObjectFactory: ObjectFactory {
 
 // MARK: ValueFactory
 
-protocol ValueFactory: RealmCollectionValue, _DefaultConstructible {
+protocol ValueFactory where Self: RealmCollectionValue {
     associatedtype Wrapped: RealmCollectionValue = Self
-    associatedtype AverageType: AddableType = Double
 
     static func values() -> [Self]
-    static func doubleValue(_ value: AverageType) -> Double
-    static func doubleValue(t value: Self) -> Double
-    static func doubleValue(w value: Wrapped) -> Double
+    static func min() -> Self
+    static func max() -> Self
 }
 extension ValueFactory {
-    static func doubleValue(_ value: Double) -> Double {
-        return value
-    }
-    static func doubleValue(t value: Self) -> Double {
-        return (value as! NSNumber).doubleValue
-    }
-    static func doubleValue(w value: Wrapped) -> Double {
-        return (value as! NSNumber).doubleValue
-    }
+    static func min() -> Self { values().first! }
+    static func max() -> Self { values().last! }
 }
 extension ValueFactory where Self: PersistableEnum {
     static func values() -> [Self] {
@@ -73,64 +64,103 @@ extension ValueFactory where Self: PersistableEnum {
     }
 }
 
+protocol NumericValueFactory: ValueFactory {
+    associatedtype AverageType: RealmCollectionValue = Double where AverageType.PersistedType: AddableType
+    static func sum(_ values: [Self]) -> Double
+    static func average() -> Double
+    static func doubleValue(_ value: Self) -> Double
+    static var zero: Self { get }
+}
+extension NumericValueFactory {
+    static func sum() -> Double {
+        return sum(values())
+    }
+    static func average() -> Double {
+        return sum() / 3
+    }
+}
+extension NumericValueFactory where Self: AdditiveArithmetic {
+    static func sum(_ values: [Self]) -> Double {
+        return doubleValue(values.reduce(.zero, +))
+    }
+}
+extension NumericValueFactory where PersistedType == Self {
+    static func doubleValue(_ value: Self) -> Double {
+        return (value as! NSNumber).doubleValue
+    }
+}
+extension NumericValueFactory where Self: PersistableEnum {
+    static func doubleValue(_ value: Self) -> Double {
+        return (value as! NSNumber).doubleValue
+    }
+}
+
 protocol ListValueFactory: ValueFactory {
     associatedtype ListRoot: Object
     static var array: KeyPath<ListRoot, List<Self>> { get }
-    static var optArray: KeyPath<ListRoot, List<Self?>> { get }
 }
 
 protocol SetValueFactory: ValueFactory {
     associatedtype SetRoot: Object
     static var mutableSet: KeyPath<SetRoot, MutableSet<Self>> { get }
-    static var optMutableSet: KeyPath<SetRoot, MutableSet<Self?>> { get }
 }
 
 protocol MapValueFactory: ValueFactory {
     associatedtype MapRoot: Object
     static var map: KeyPath<MapRoot, Map<String, Self>> { get }
+}
+
+protocol ListValueFactoryOptional: ListValueFactory where Self: _RealmCollectionValueInsideOptional {
+    static var optArray: KeyPath<ListRoot, List<Self?>> { get }
+}
+
+protocol SetValueFactoryOptional: SetValueFactory where Self: _RealmCollectionValueInsideOptional {
+    static var optMutableSet: KeyPath<SetRoot, MutableSet<Self?>> { get }
+}
+
+protocol MapValueFactoryOptional: MapValueFactory where Self: _RealmCollectionValueInsideOptional {
     static var optMap: KeyPath<MapRoot, Map<String, Self?>> { get }
 }
 
 // MARK: Optional
 
-extension Optional: _DefaultConstructible where Wrapped: _DefaultConstructible {
-    public init() {
-        self = .none
-    }
-}
-
-extension Optional: ValueFactory where Wrapped: ValueFactory & _DefaultConstructible {
-    typealias AverageType = Wrapped.AverageType
-
-    static func doubleValue(_ value: Wrapped.AverageType) -> Double {
-        Wrapped.doubleValue(value)
-    }
-    static func doubleValue(t value: Self) -> Double {
-        return Wrapped.doubleValue(t: value!)
-    }
-    static func doubleValue(w value: Wrapped) -> Double {
-        return Wrapped.doubleValue(w: value as! Wrapped.Wrapped)
-    }
-
+extension Optional: ValueFactory where Wrapped: ValueFactory & _RealmCollectionValueInsideOptional {
     static func values() -> [Self] {
         var v = Array<Self>(Wrapped.values())
         v.remove(at: 1)
         v.insert(nil, at: 0)
         return v
     }
+
+    static func min() -> Self { Self.some(Wrapped.min()) }
 }
 
-extension Optional: ListValueFactory where Wrapped: ListValueFactory {
+extension Optional: NumericValueFactory where Wrapped: NumericValueFactory & _RealmCollectionValueInsideOptional {
+    typealias AverageType = Wrapped.AverageType
+    static func doubleValue(_ value: Self) -> Double {
+        return Wrapped.doubleValue(value!)
+    }
+
+    static func min() -> Self { Self.some(Wrapped.min()) }
+    static func sum(_ values: [Self]) -> Double {
+        return Wrapped.sum(values.compactMap { $0 })
+    }
+    static func average() -> Double {
+        return sum() / 2
+    }
+    static var zero: Optional<Wrapped> {
+        Wrapped.zero
+    }
+}
+
+extension Optional: ListValueFactory where Wrapped: ListValueFactoryOptional {
     static var array: KeyPath<Wrapped.ListRoot, List<Self>> { Wrapped.optArray }
-    static var optArray: KeyPath<Wrapped.ListRoot, List<Self?>> { fatalError() }
 }
-extension Optional: SetValueFactory where Wrapped: SetValueFactory {
+extension Optional: SetValueFactory where Wrapped: SetValueFactoryOptional {
     static var mutableSet: KeyPath<Wrapped.SetRoot, MutableSet<Self>> { Wrapped.optMutableSet }
-    static var optMutableSet: KeyPath<Wrapped.SetRoot, MutableSet<Self?>> { fatalError() }
 }
-extension Optional: MapValueFactory where Wrapped: MapValueFactory {
+extension Optional: MapValueFactory where Wrapped: MapValueFactoryOptional {
     static var map: KeyPath<Wrapped.MapRoot, Map<String, Self>> { Wrapped.optMap }
-    static var optMap: KeyPath<Wrapped.MapRoot, Map<String, Self?>> { fatalError() }
 }
 
 // MARK: - Bool
@@ -141,162 +171,162 @@ extension Bool: ValueFactory {
         return _values
     }
 }
-extension Bool: ListValueFactory {
+extension Bool: ListValueFactoryOptional {
     static var array: KeyPath<ModernAllTypesObject, List<Bool>> { \.arrayBool }
     static var optArray: KeyPath<ModernAllTypesObject, List<Bool?>> { \.arrayOptBool }
 }
-extension Bool: SetValueFactory {
+extension Bool: SetValueFactoryOptional {
     static var mutableSet: KeyPath<ModernAllTypesObject, MutableSet<Bool>> { \.setBool }
     static var optMutableSet: KeyPath<ModernAllTypesObject, MutableSet<Bool?>> { \.setOptBool }
 }
-extension Bool: MapValueFactory {
+extension Bool: MapValueFactoryOptional {
     static var map: KeyPath<ModernAllTypesObject, Map<String, Bool>> { \.mapBool }
     static var optMap: KeyPath<ModernAllTypesObject, Map<String, Bool?>> { \.mapOptBool }
 }
 
 // MARK: - Int
 
-extension Int: ValueFactory {
+extension Int: NumericValueFactory {
     private static let _values: [Int] = [1, 2, 3]
     static func values() -> [Int] {
         return _values
     }
 }
-extension Int: ListValueFactory {
+extension Int: ListValueFactoryOptional {
     static var array: KeyPath<ModernAllTypesObject, List<Int>> { \.arrayInt }
     static var optArray: KeyPath<ModernAllTypesObject, List<Int?>> { \.arrayOptInt }
 }
-extension Int: SetValueFactory {
+extension Int: SetValueFactoryOptional {
     static var mutableSet: KeyPath<ModernAllTypesObject, MutableSet<Int>> { \.setInt }
     static var optMutableSet: KeyPath<ModernAllTypesObject, MutableSet<Int?>> { \.setOptInt }
 }
-extension Int: MapValueFactory {
+extension Int: MapValueFactoryOptional {
     static var map: KeyPath<ModernAllTypesObject, Map<String, Int>> { \.mapInt }
     static var optMap: KeyPath<ModernAllTypesObject, Map<String, Int?>> { \.mapOptInt }
 }
 
 // MARK: - Int8
 
-extension Int8: ValueFactory {
+extension Int8: NumericValueFactory {
     private static let _values: [Int8] = [1, 2, 3]
     static func values() -> [Int8] {
         return _values
     }
 }
-extension Int8: ListValueFactory {
+extension Int8: ListValueFactoryOptional {
     static var array: KeyPath<ModernAllTypesObject, List<Int8>> { \.arrayInt8 }
     static var optArray: KeyPath<ModernAllTypesObject, List<Int8?>> { \.arrayOptInt8 }
 }
-extension Int8: SetValueFactory {
+extension Int8: SetValueFactoryOptional {
     static var mutableSet: KeyPath<ModernAllTypesObject, MutableSet<Int8>> { \.setInt8 }
     static var optMutableSet: KeyPath<ModernAllTypesObject, MutableSet<Int8?>> { \.setOptInt8 }
 }
-extension Int8: MapValueFactory {
+extension Int8: MapValueFactoryOptional {
     static var map: KeyPath<ModernAllTypesObject, Map<String, Int8>> { \.mapInt8 }
     static var optMap: KeyPath<ModernAllTypesObject, Map<String, Int8?>> { \.mapOptInt8 }
 }
 
 // MARK: - Int16
 
-extension Int16: ValueFactory {
+extension Int16: NumericValueFactory {
     private static let _values: [Int16] = [1, 2, 3]
     static func values() -> [Int16] {
         return _values
     }
 }
-extension Int16: ListValueFactory {
+extension Int16: ListValueFactoryOptional {
     static var array: KeyPath<ModernAllTypesObject, List<Int16>> { \.arrayInt16 }
     static var optArray: KeyPath<ModernAllTypesObject, List<Int16?>> { \.arrayOptInt16 }
 }
-extension Int16: SetValueFactory {
+extension Int16: SetValueFactoryOptional {
     static var mutableSet: KeyPath<ModernAllTypesObject, MutableSet<Int16>> { \.setInt16 }
     static var optMutableSet: KeyPath<ModernAllTypesObject, MutableSet<Int16?>> { \.setOptInt16 }
 }
-extension Int16: MapValueFactory {
+extension Int16: MapValueFactoryOptional {
     static var map: KeyPath<ModernAllTypesObject, Map<String, Int16>> { \.mapInt16 }
     static var optMap: KeyPath<ModernAllTypesObject, Map<String, Int16?>> { \.mapOptInt16 }
 }
 
 // MARK: - Int32
 
-extension Int32: ValueFactory {
+extension Int32: NumericValueFactory {
     private static let _values: [Int32] = [1, 2, 3]
     static func values() -> [Int32] {
         return _values
     }
 }
-extension Int32: ListValueFactory {
+extension Int32: ListValueFactoryOptional {
     static var array: KeyPath<ModernAllTypesObject, List<Int32>> { \.arrayInt32 }
     static var optArray: KeyPath<ModernAllTypesObject, List<Int32?>> { \.arrayOptInt32 }
 }
-extension Int32: SetValueFactory {
+extension Int32: SetValueFactoryOptional {
     static var mutableSet: KeyPath<ModernAllTypesObject, MutableSet<Int32>> { \.setInt32 }
     static var optMutableSet: KeyPath<ModernAllTypesObject, MutableSet<Int32?>> { \.setOptInt32 }
 }
-extension Int32: MapValueFactory {
+extension Int32: MapValueFactoryOptional {
     static var map: KeyPath<ModernAllTypesObject, Map<String, Int32>> { \.mapInt32 }
     static var optMap: KeyPath<ModernAllTypesObject, Map<String, Int32?>> { \.mapOptInt32 }
 }
 
 // MARK: - Int64
 
-extension Int64: ValueFactory {
+extension Int64: NumericValueFactory {
     private static let _values: [Int64] = [1, 2, 3]
     static func values() -> [Int64] {
         return _values
     }
 }
-extension Int64: ListValueFactory {
+extension Int64: ListValueFactoryOptional {
     static var array: KeyPath<ModernAllTypesObject, List<Int64>> { \.arrayInt64 }
     static var optArray: KeyPath<ModernAllTypesObject, List<Int64?>> { \.arrayOptInt64 }
 }
-extension Int64: SetValueFactory {
+extension Int64: SetValueFactoryOptional {
     static var mutableSet: KeyPath<ModernAllTypesObject, MutableSet<Int64>> { \.setInt64 }
     static var optMutableSet: KeyPath<ModernAllTypesObject, MutableSet<Int64?>> { \.setOptInt64 }
 }
-extension Int64: MapValueFactory {
+extension Int64: MapValueFactoryOptional {
     static var map: KeyPath<ModernAllTypesObject, Map<String, Int64>> { \.mapInt64 }
     static var optMap: KeyPath<ModernAllTypesObject, Map<String, Int64?>> { \.mapOptInt64 }
 }
 
 // MARK: - Float
 
-extension Float: ValueFactory {
+extension Float: NumericValueFactory {
     private static let _values: [Float] = [1.1, 2.2, 3.3]
     static func values() -> [Float] {
         return _values
     }
 }
-extension Float: ListValueFactory {
+extension Float: ListValueFactoryOptional {
     static var array: KeyPath<ModernAllTypesObject, List<Float>> { \.arrayFloat }
     static var optArray: KeyPath<ModernAllTypesObject, List<Float?>> { \.arrayOptFloat }
 }
-extension Float: SetValueFactory {
+extension Float: SetValueFactoryOptional {
     static var mutableSet: KeyPath<ModernAllTypesObject, MutableSet<Float>> { \.setFloat }
     static var optMutableSet: KeyPath<ModernAllTypesObject, MutableSet<Float?>> { \.setOptFloat }
 }
-extension Float: MapValueFactory {
+extension Float: MapValueFactoryOptional {
     static var map: KeyPath<ModernAllTypesObject, Map<String, Float>> { \.mapFloat }
     static var optMap: KeyPath<ModernAllTypesObject, Map<String, Float?>> { \.mapOptFloat }
 }
 
 // MARK: - Double
 
-extension Double: ValueFactory {
+extension Double: NumericValueFactory {
     private static let _values: [Double] = [1.1, 2.2, 3.3]
     static func values() -> [Double] {
         return _values
     }
 }
-extension Double: ListValueFactory {
+extension Double: ListValueFactoryOptional {
     static var array: KeyPath<ModernAllTypesObject, List<Double>> { \.arrayDouble }
     static var optArray: KeyPath<ModernAllTypesObject, List<Double?>> { \.arrayOptDouble }
 }
-extension Double: SetValueFactory {
+extension Double: SetValueFactoryOptional {
     static var mutableSet: KeyPath<ModernAllTypesObject, MutableSet<Double>> { \.setDouble }
     static var optMutableSet: KeyPath<ModernAllTypesObject, MutableSet<Double?>> { \.setOptDouble }
 }
-extension Double: MapValueFactory {
+extension Double: MapValueFactoryOptional {
     static var map: KeyPath<ModernAllTypesObject, Map<String, Double>> { \.mapDouble }
     static var optMap: KeyPath<ModernAllTypesObject, Map<String, Double?>> { \.mapOptDouble }
 }
@@ -309,15 +339,15 @@ extension String: ValueFactory {
         return _values
     }
 }
-extension String: ListValueFactory {
+extension String: ListValueFactoryOptional {
     static var array: KeyPath<ModernAllTypesObject, List<String>> { \.arrayString }
     static var optArray: KeyPath<ModernAllTypesObject, List<String?>> { \.arrayOptString }
 }
-extension String: SetValueFactory {
+extension String: SetValueFactoryOptional {
     static var mutableSet: KeyPath<ModernAllTypesObject, MutableSet<String>> { \.setString }
     static var optMutableSet: KeyPath<ModernAllTypesObject, MutableSet<String?>> { \.setOptString }
 }
-extension String: MapValueFactory {
+extension String: MapValueFactoryOptional {
     static var map: KeyPath<ModernAllTypesObject, Map<String, String>> { \.mapString }
     static var optMap: KeyPath<ModernAllTypesObject, Map<String, String?>> { \.mapOptString }
 }
@@ -330,15 +360,15 @@ extension Data: ValueFactory {
         return _values
     }
 }
-extension Data: ListValueFactory {
+extension Data: ListValueFactoryOptional {
     static var array: KeyPath<ModernAllTypesObject, List<Data>> { \.arrayBinary }
     static var optArray: KeyPath<ModernAllTypesObject, List<Data?>> { \.arrayOptBinary }
 }
-extension Data: SetValueFactory {
+extension Data: SetValueFactoryOptional {
     static var mutableSet: KeyPath<ModernAllTypesObject, MutableSet<Data>> { \.setBinary }
     static var optMutableSet: KeyPath<ModernAllTypesObject, MutableSet<Data?>> { \.setOptBinary }
 }
-extension Data: MapValueFactory {
+extension Data: MapValueFactoryOptional {
     static var map: KeyPath<ModernAllTypesObject, Map<String, Data>> { \.mapBinary }
     static var optMap: KeyPath<ModernAllTypesObject, Map<String, Data?>> { \.mapOptBinary }
 }
@@ -351,22 +381,22 @@ extension Date: ValueFactory {
         return _values
     }
 }
-extension Date: ListValueFactory {
+extension Date: ListValueFactoryOptional {
     static var array: KeyPath<ModernAllTypesObject, List<Date>> { \.arrayDate }
     static var optArray: KeyPath<ModernAllTypesObject, List<Date?>> { \.arrayOptDate }
 }
-extension Date: SetValueFactory {
+extension Date: SetValueFactoryOptional {
     static var mutableSet: KeyPath<ModernAllTypesObject, MutableSet<Date>> { \.setDate }
     static var optMutableSet: KeyPath<ModernAllTypesObject, MutableSet<Date?>> { \.setOptDate }
 }
-extension Date: MapValueFactory {
+extension Date: MapValueFactoryOptional {
     static var map: KeyPath<ModernAllTypesObject, Map<String, Date>> { \.mapDate }
     static var optMap: KeyPath<ModernAllTypesObject, Map<String, Date?>> { \.mapOptDate }
 }
 
 // MARK: - Decimal128
 
-extension Decimal128: ValueFactory {
+extension Decimal128: NumericValueFactory {
     private static let _values: [Decimal128] = [Decimal128(number: 1), Decimal128(number: 2), Decimal128(number: 3)]
     static func values() -> [Decimal128] {
         return _values
@@ -375,22 +405,16 @@ extension Decimal128: ValueFactory {
     static func doubleValue(_ value: Decimal128) -> Double {
         return value.doubleValue
     }
-    static func doubleValue(t value: Decimal128) -> Double {
-        return value.doubleValue
-    }
-    static func doubleValue(w value: Decimal128) -> Double {
-        return value.doubleValue
-    }
 }
-extension Decimal128: ListValueFactory {
+extension Decimal128: ListValueFactoryOptional {
     static var array: KeyPath<ModernAllTypesObject, List<Decimal128>> { \.arrayDecimal }
     static var optArray: KeyPath<ModernAllTypesObject, List<Decimal128?>> { \.arrayOptDecimal }
 }
-extension Decimal128: SetValueFactory {
+extension Decimal128: SetValueFactoryOptional {
     static var mutableSet: KeyPath<ModernAllTypesObject, MutableSet<Decimal128>> { \.setDecimal }
     static var optMutableSet: KeyPath<ModernAllTypesObject, MutableSet<Decimal128?>> { \.setOptDecimal }
 }
-extension Decimal128: MapValueFactory {
+extension Decimal128: MapValueFactoryOptional {
     static var map: KeyPath<ModernAllTypesObject, Map<String, Decimal128>> { \.mapDecimal }
     static var optMap: KeyPath<ModernAllTypesObject, Map<String, Decimal128?>> { \.mapOptDecimal }
 }
@@ -403,15 +427,15 @@ extension ObjectId: ValueFactory {
         return _values
     }
 }
-extension ObjectId: ListValueFactory {
+extension ObjectId: ListValueFactoryOptional {
     static var array: KeyPath<ModernAllTypesObject, List<ObjectId>> { \.arrayObjectId }
     static var optArray: KeyPath<ModernAllTypesObject, List<ObjectId?>> { \.arrayOptObjectId }
 }
-extension ObjectId: SetValueFactory {
+extension ObjectId: SetValueFactoryOptional {
     static var mutableSet: KeyPath<ModernAllTypesObject, MutableSet<ObjectId>> { \.setObjectId }
     static var optMutableSet: KeyPath<ModernAllTypesObject, MutableSet<ObjectId?>> { \.setOptObjectId }
 }
-extension ObjectId: MapValueFactory {
+extension ObjectId: MapValueFactoryOptional {
     static var map: KeyPath<ModernAllTypesObject, Map<String, ObjectId>> { \.mapObjectId }
     static var optMap: KeyPath<ModernAllTypesObject, Map<String, ObjectId?>> { \.mapOptObjectId }
 }
@@ -424,15 +448,15 @@ extension UUID: ValueFactory {
         return _values
     }
 }
-extension UUID: ListValueFactory {
+extension UUID: ListValueFactoryOptional {
     static var array: KeyPath<ModernAllTypesObject, List<UUID>> { \.arrayUuid }
     static var optArray: KeyPath<ModernAllTypesObject, List<UUID?>> { \.arrayOptUuid }
 }
-extension UUID: SetValueFactory {
+extension UUID: SetValueFactoryOptional {
     static var mutableSet: KeyPath<ModernAllTypesObject, MutableSet<UUID>> { \.setUuid }
     static var optMutableSet: KeyPath<ModernAllTypesObject, MutableSet<UUID?>> { \.setOptUuid }
 }
-extension UUID: MapValueFactory {
+extension UUID: MapValueFactoryOptional {
     static var map: KeyPath<ModernAllTypesObject, Map<String, UUID>> { \.mapUuid }
     static var optMap: KeyPath<ModernAllTypesObject, Map<String, UUID?>> { \.mapOptUuid }
 }
@@ -449,15 +473,15 @@ extension EnumInt: ValueFactory {
         return EnumInt.allCases
     }
 }
-extension EnumInt: ListValueFactory {
+extension EnumInt: ListValueFactoryOptional {
     static var array: KeyPath<ModernCollectionsOfEnums, List<EnumInt>> { \.listInt }
     static var optArray: KeyPath<ModernCollectionsOfEnums, List<EnumInt?>> { \.listIntOpt }
 }
-extension EnumInt: SetValueFactory {
+extension EnumInt: SetValueFactoryOptional {
     static var mutableSet: KeyPath<ModernCollectionsOfEnums, MutableSet<EnumInt>> { \.setInt }
     static var optMutableSet: KeyPath<ModernCollectionsOfEnums, MutableSet<EnumInt?>> { \.setIntOpt }
 }
-extension EnumInt: MapValueFactory {
+extension EnumInt: MapValueFactoryOptional {
     static var map: KeyPath<ModernCollectionsOfEnums, Map<String, EnumInt>> { \.mapInt }
     static var optMap: KeyPath<ModernCollectionsOfEnums, Map<String, EnumInt?>> { \.mapIntOpt }
 }
@@ -474,15 +498,15 @@ extension EnumInt8: ValueFactory {
         return EnumInt8.allCases
     }
 }
-extension EnumInt8: ListValueFactory {
+extension EnumInt8: ListValueFactoryOptional {
     static var array: KeyPath<ModernCollectionsOfEnums, List<EnumInt8>> { \.listInt8 }
     static var optArray: KeyPath<ModernCollectionsOfEnums, List<EnumInt8?>> { \.listInt8Opt }
 }
-extension EnumInt8: SetValueFactory {
+extension EnumInt8: SetValueFactoryOptional {
     static var mutableSet: KeyPath<ModernCollectionsOfEnums, MutableSet<EnumInt8>> { \.setInt8 }
     static var optMutableSet: KeyPath<ModernCollectionsOfEnums, MutableSet<EnumInt8?>> { \.setInt8Opt }
 }
-extension EnumInt8: MapValueFactory {
+extension EnumInt8: MapValueFactoryOptional {
     static var map: KeyPath<ModernCollectionsOfEnums, Map<String, EnumInt8>> { \.mapInt8 }
     static var optMap: KeyPath<ModernCollectionsOfEnums, Map<String, EnumInt8?>> { \.mapInt8Opt }
 }
@@ -499,15 +523,15 @@ extension EnumInt16: ValueFactory {
         return EnumInt16.allCases
     }
 }
-extension EnumInt16: ListValueFactory {
+extension EnumInt16: ListValueFactoryOptional {
     static var array: KeyPath<ModernCollectionsOfEnums, List<EnumInt16>> { \.listInt16 }
     static var optArray: KeyPath<ModernCollectionsOfEnums, List<EnumInt16?>> { \.listInt16Opt }
 }
-extension EnumInt16: SetValueFactory {
+extension EnumInt16: SetValueFactoryOptional {
     static var mutableSet: KeyPath<ModernCollectionsOfEnums, MutableSet<EnumInt16>> { \.setInt16 }
     static var optMutableSet: KeyPath<ModernCollectionsOfEnums, MutableSet<EnumInt16?>> { \.setInt16Opt }
 }
-extension EnumInt16: MapValueFactory {
+extension EnumInt16: MapValueFactoryOptional {
     static var map: KeyPath<ModernCollectionsOfEnums, Map<String, EnumInt16>> { \.mapInt16 }
     static var optMap: KeyPath<ModernCollectionsOfEnums, Map<String, EnumInt16?>> { \.mapInt16Opt }
 }
@@ -524,15 +548,15 @@ extension EnumInt32: ValueFactory {
         return EnumInt32.allCases
     }
 }
-extension EnumInt32: ListValueFactory {
+extension EnumInt32: ListValueFactoryOptional {
     static var array: KeyPath<ModernCollectionsOfEnums, List<EnumInt32>> { \.listInt32 }
     static var optArray: KeyPath<ModernCollectionsOfEnums, List<EnumInt32?>> { \.listInt32Opt }
 }
-extension EnumInt32: SetValueFactory {
+extension EnumInt32: SetValueFactoryOptional {
     static var mutableSet: KeyPath<ModernCollectionsOfEnums, MutableSet<EnumInt32>> { \.setInt32 }
     static var optMutableSet: KeyPath<ModernCollectionsOfEnums, MutableSet<EnumInt32?>> { \.setInt32Opt }
 }
-extension EnumInt32: MapValueFactory {
+extension EnumInt32: MapValueFactoryOptional {
     static var map: KeyPath<ModernCollectionsOfEnums, Map<String, EnumInt32>> { \.mapInt32 }
     static var optMap: KeyPath<ModernCollectionsOfEnums, Map<String, EnumInt32?>> { \.mapInt32Opt }
 }
@@ -549,15 +573,15 @@ extension EnumInt64: ValueFactory {
         return EnumInt64.allCases
     }
 }
-extension EnumInt64: ListValueFactory {
+extension EnumInt64: ListValueFactoryOptional {
     static var array: KeyPath<ModernCollectionsOfEnums, List<EnumInt64>> { \.listInt64 }
     static var optArray: KeyPath<ModernCollectionsOfEnums, List<EnumInt64?>> { \.listInt64Opt }
 }
-extension EnumInt64: SetValueFactory {
+extension EnumInt64: SetValueFactoryOptional {
     static var mutableSet: KeyPath<ModernCollectionsOfEnums, MutableSet<EnumInt64>> { \.setInt64 }
     static var optMutableSet: KeyPath<ModernCollectionsOfEnums, MutableSet<EnumInt64?>> { \.setInt64Opt }
 }
-extension EnumInt64: MapValueFactory {
+extension EnumInt64: MapValueFactoryOptional {
     static var map: KeyPath<ModernCollectionsOfEnums, Map<String, EnumInt64>> { \.mapInt64 }
     static var optMap: KeyPath<ModernCollectionsOfEnums, Map<String, EnumInt64?>> { \.mapInt64Opt }
 }
@@ -574,15 +598,15 @@ extension EnumFloat: ValueFactory {
         return EnumFloat.allCases
     }
 }
-extension EnumFloat: ListValueFactory {
+extension EnumFloat: ListValueFactoryOptional {
     static var array: KeyPath<ModernCollectionsOfEnums, List<EnumFloat>> { \.listFloat }
     static var optArray: KeyPath<ModernCollectionsOfEnums, List<EnumFloat?>> { \.listFloatOpt }
 }
-extension EnumFloat: SetValueFactory {
+extension EnumFloat: SetValueFactoryOptional {
     static var mutableSet: KeyPath<ModernCollectionsOfEnums, MutableSet<EnumFloat>> { \.setFloat }
     static var optMutableSet: KeyPath<ModernCollectionsOfEnums, MutableSet<EnumFloat?>> { \.setFloatOpt }
 }
-extension EnumFloat: MapValueFactory {
+extension EnumFloat: MapValueFactoryOptional {
     static var map: KeyPath<ModernCollectionsOfEnums, Map<String, EnumFloat>> { \.mapFloat }
     static var optMap: KeyPath<ModernCollectionsOfEnums, Map<String, EnumFloat?>> { \.mapFloatOpt }
 }
@@ -599,15 +623,15 @@ extension EnumDouble: ValueFactory {
         return EnumDouble.allCases
     }
 }
-extension EnumDouble: ListValueFactory {
+extension EnumDouble: ListValueFactoryOptional {
     static var array: KeyPath<ModernCollectionsOfEnums, List<EnumDouble>> { \.listDouble }
     static var optArray: KeyPath<ModernCollectionsOfEnums, List<EnumDouble?>> { \.listDoubleOpt }
 }
-extension EnumDouble: SetValueFactory {
+extension EnumDouble: SetValueFactoryOptional {
     static var mutableSet: KeyPath<ModernCollectionsOfEnums, MutableSet<EnumDouble>> { \.setDouble }
     static var optMutableSet: KeyPath<ModernCollectionsOfEnums, MutableSet<EnumDouble?>> { \.setDoubleOpt }
 }
-extension EnumDouble: MapValueFactory {
+extension EnumDouble: MapValueFactoryOptional {
     static var map: KeyPath<ModernCollectionsOfEnums, Map<String, EnumDouble>> { \.mapDouble }
     static var optMap: KeyPath<ModernCollectionsOfEnums, Map<String, EnumDouble?>> { \.mapDoubleOpt }
 }
@@ -624,15 +648,15 @@ extension EnumString: ValueFactory {
         return EnumString.allCases
     }
 }
-extension EnumString: ListValueFactory {
+extension EnumString: ListValueFactoryOptional {
     static var array: KeyPath<ModernCollectionsOfEnums, List<EnumString>> { \.listString }
     static var optArray: KeyPath<ModernCollectionsOfEnums, List<EnumString?>> { \.listStringOpt }
 }
-extension EnumString: SetValueFactory {
+extension EnumString: SetValueFactoryOptional {
     static var mutableSet: KeyPath<ModernCollectionsOfEnums, MutableSet<EnumString>> { \.setString }
     static var optMutableSet: KeyPath<ModernCollectionsOfEnums, MutableSet<EnumString?>> { \.setStringOpt }
 }
-extension EnumString: MapValueFactory {
+extension EnumString: MapValueFactoryOptional {
     static var map: KeyPath<ModernCollectionsOfEnums, Map<String, EnumString>> { \.mapString }
     static var optMap: KeyPath<ModernCollectionsOfEnums, Map<String, EnumString?>> { \.mapStringOpt }
 }